refactor(126181): Separando os grupos por arquivo
